mcps/obsidian/pyproject.toml
[project]
name = "obsidian-mcp-server"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp==3.1.1",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "urllib3>=2.0",
    "pypdf>=4.0",
    "pyyaml>=6.0",
]
